+++ src/Credentials/GCECredentials.php
@@ -24,7 +24,6 @@ use Google\Auth\HttpHandler\HttpHandlerFactory;
 use Google\Auth\Iam;
 use Google\Auth\ProjectIdProviderInterface;
 use Google\Auth\SignBlobInterface;
-use Google\Auth\UpdateMetadataInterface;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Exception\RequestException;
@@ -59,8 +58,7 @@ use InvalidArgumentException;
 class GCECredentials extends CredentialsLoader implements
     SignBlobInterface,
     ProjectIdProviderInterface,
-    GetQuotaProjectInterface,
-    UpdateMetadataInterface
+    GetQuotaProjectInterface
 {
     // phpcs:disable
     const cacheKey = 'GOOGLE_AUTH_PHP_GCE';
