+++ src/CredentialsLoader.php
@@ -35,8 +35,6 @@ abstract class CredentialsLoader implements
     const WELL_KNOWN_PATH = 'gcloud/application_default_credentials.json';
     const NON_WINDOWS_WELL_KNOWN_PATH_BASE = '.config';
 
-    use UpdateMetadataTrait;
-
     /**
      * @param string $cause
      * @return string
@@ -212,5 +210,32 @@ abstract class CredentialsLoader implements
     public function getUpdateMetadataFunc()
     {
         return array($this, 'updateMetadata');
+    }
+
+    /**
+     * Updates metadata with the authorization token.
+     *
+     * @param array $metadata metadata hashmap
+     * @param string $authUri optional auth uri
+     * @param callable $httpHandler callback which delivers psr7 request
+     * @return array updated metadata hashmap
+     */
+    public function updateMetadata(
+        $metadata,
+        $authUri = null,
+        callable $httpHandler = null
+    ) {
+        if (isset($metadata[self::AUTH_METADATA_KEY])) {
+            // Auth metadata has already been set
+            return $metdadata;
+        }
+        $result = $this->fetchAuthToken($httpHandler);
+        if (!isset($result['access_token'])) {
+            return $metadata;
+        }
+        $metadata_copy = $metadata;
+        $metadata_copy[self::AUTH_METADATA_KEY] = array('Bearer ' . $result['access_token']);
+
+        return $metadata_copy;
     }
 }
